refactor: First pass at splitting up ResumePage
