--- apps/editor.planx.uk/src/pages/Preview/ResumePage.tsx
@@ -1,334 +1,0 @@
-import Box from "@mui/material/Box";
-import Link from "@mui/material/Link";
-import Typography from "@mui/material/Typography";
-import { PaymentRequest } from "@opensystemslab/planx-core/types";
-import Card from "@planx/components/shared/Preview/Card";
-import { CardHeader } from "@planx/components/shared/Preview/CardHeader/CardHeader";
-import { useMutation } from "@tanstack/react-query";
-import { sendResumeEmail, validateSession } from "api/saveAndReturn/requests";
-import { ReconciliationResponse } from "api/saveAndReturn/types";
-import { isAxiosError } from "axios";
-import DelayedLoadingIndicator from "components/DelayedLoadingIndicator/DelayedLoadingIndicator";
-import { useFormik } from "formik";
-import { useStore } from "pages/FlowEditor/lib/store";
-import React, { useEffect, useState } from "react";
-import { Link as ReactNaviLink, useCurrentRoute } from "react-navi";
-import { ApplicationPath } from "types";
-import InputLabel from "ui/public/InputLabel";
-import Input from "ui/shared/Input/Input";
-import InputRow from "ui/shared/InputRow";
-import { object, string } from "yup";
-
-import ReconciliationPage from "./ReconciliationPage";
-import StatusPage from "./StatusPage";
-
-enum Status {
-  EmailRequired,
-  Validating,
-  Validated,
-  InvalidSession,
-  Success,
-  Error,
-  LockedSession,
-}
-// A minimal representation of a PaymentRequest for display purposes
-type MinPaymentRequest = Pick<
-  PaymentRequest,
-  "id" | "payeeEmail" | "payeeName"
->;
-
-interface LockedSessionResponse {
-  paymentRequest: MinPaymentRequest;
-  status: number;
-  message: string;
-}
-
-export const EmailRequired: React.FC<{ setEmail: (email: string) => void }> = ({
-  setEmail,
-}) => {
-  const emailSchema = object({
-    email: string().email("Invalid email").required("Email address required"),
-  });
-
-  const formik = useFormik({
-    initialValues: {
-      email: "",
-    },
-    onSubmit: (values) => setEmail(values.email),
-    validateOnChange: false,
-    validateOnBlur: false,
-    validationSchema: emailSchema,
-  });
-
-  return (
-    <Box width="100%">
-      <Card handleSubmit={formik.handleSubmit}>
-        <CardHeader
-          title="Resume your application"
-          description="Enter your email to resume your application."
-        ></CardHeader>
-        <InputRow>
-          <InputLabel label={"Email address"} htmlFor={"email"}>
-            <Input
-              bordered
-              errorMessage={
-                formik.touched.email && formik.errors.email
-                  ? formik.errors.email
-                  : undefined
-              }
-              id="email"
-              name="email"
-              onBlur={formik.handleBlur}
-              onChange={formik.handleChange}
-              type="email"
-              autoComplete="email"
-              value={formik.values.email}
-            ></Input>
-          </InputLabel>
-        </InputRow>
-      </Card>
-    </Box>
-  );
-};
-
-export const EmailError: React.FC<{ retry: () => void }> = ({ retry }) => {
-  return (
-    <StatusPage
-      bannerHeading="Email not sent"
-      buttonText="Retry"
-      onButtonClick={retry}
-    >
-      <Typography variant="body1">
-        We are having trouble sending emails at the moment.
-        <br />
-        <br />
-        We have saved your application. We will try to send the email again
-        later.
-      </Typography>
-    </StatusPage>
-  );
-};
-
-export const EmailSuccess: React.FC = () => {
-  return (
-    <StatusPage bannerHeading="Check your email">
-      <Typography variant="body1">
-        If you have any draft applications we have sent you an email that
-        contains a link. Use this link to access your applications.
-        <br />
-        <br />
-        You may now close this tab.
-      </Typography>
-    </StatusPage>
-  );
-};
-
-export const ValidationSuccess: React.FC<{
-  reconciliationResponse: ReconciliationResponse;
-  continueApplication: () => void;
-}> = ({ reconciliationResponse, continueApplication }) => {
-  return (
-    <ReconciliationPage
-      bannerHeading="Resume your application"
-      reconciliationResponse={reconciliationResponse}
-      buttonText="Continue"
-      onButtonClick={() => continueApplication()}
-    ></ReconciliationPage>
-  );
-};
-
-export const InvalidSession: React.FC<{
-  retry: () => void;
-}> = ({ retry }) => (
-  <StatusPage
-    bannerHeading="We can't find your application"
-    buttonText="Try again"
-    onButtonClick={retry}
-    additionalOption="startNewApplication"
-  >
-    <Typography variant="h3" component="h2">
-      Reasons
-    </Typography>
-    <Typography variant="body1">
-      <br />
-      This may be because your application has expired or there was a mistake in
-      your email address.
-      <br />
-      <br />
-      Try entering your email address again.
-    </Typography>
-  </StatusPage>
-);
-
-export const LockedSession: React.FC<{
-  paymentRequest?: MinPaymentRequest;
-}> = ({ paymentRequest }) => (
-  <StatusPage
-    bannerHeading="Sorry, you can't make changes to this application"
-    additionalOption="startNewApplication"
-  >
-    <Typography variant="body1">
-      This is because you've invited {paymentRequest?.payeeName} (
-      <Link href={`mailto:${paymentRequest?.payeeEmail}`}>
-        {paymentRequest?.payeeEmail}
-      </Link>
-      ) to pay for this application and changes might affect the fee.
-      <br />
-      <br />
-      You can{" "}
-      <Link
-        component={ReactNaviLink}
-        href={`../pay?paymentRequestId=${paymentRequest?.id}`}
-      >
-        pay for this application yourself on the payment page
-      </Link>
-    </Typography>
-  </StatusPage>
-);
-
-/**
- * If an email is passed in as a query param, do not prompt a user for this
- *
- * This means there's not an additional step of friction from logging into LPS,
- * or redirecting back from GOV.UK Pay
- *
- * XXX: Won't work locally as referrer is stripped from the browser when navigating from HTTPS to HTTP (localhost)
- */
-const getInitialEmailValue = (emailQueryParam?: string) => {
-  const trustedAddresses = [
-    "https://www.payments.service.gov.uk/",
-    "https://card.payments.service.gov.uk/",
-    "https://localplanning.editor.planx.dev/",
-    "https://www.localplanning.services/",
-  ];
-
-  const trustedPatterns = [
-    /^https:\/\/localplanning\.\d{4,5}\.planx\.pizza\/$/,
-  ];
-
-  const isRedirectFromTrustedSource =
-    trustedAddresses.includes(document.referrer) ||
-    trustedPatterns.some((pattern) => pattern.test(document.referrer));
-
-  if (isRedirectFromTrustedSource && emailQueryParam) return emailQueryParam;
-  return "";
-};
-
-/**
- * Component which handles the "Resume" page used for Save & Return
- * The user can access this page via four "paths"
- * 1. Directly via PlanX, user enters email to trigger "dashboard" email with resume magic links
- * 2. Magic link in email with a sessionId, user enters email to continue application
- * 3. Redirect back from GovPay - sessionId and email come from query params
- * 4. Redirect from localplanning.services - sessionId and email come from query params
- */
-const ResumePage: React.FC = () => {
-  const route = useCurrentRoute();
-
-  const [pageStatus, setPageStatus] = useState<Status>(Status.EmailRequired);
-  const [email, setEmail] = useState<string>(
-    getInitialEmailValue(decodeURIComponent(route.url.query.email)),
-  );
-  const [paymentRequest, setPaymentRequest] = useState<MinPaymentRequest>();
-  const sessionId = useCurrentRoute().url.query.sessionId;
-  const [reconciliationResponse, setReconciliationResponse] =
-    useState<ReconciliationResponse>();
-
-  useEffect(() => {
-    if (email) handleSubmit();
-  }, [email]);
-
-  const teamSlug = useStore((state) => state.teamSlug);
-  const resumeSession = useStore((state) => state.resumeSession);
-
-  /**
-   * Continue application following successful validation & reconciliation
-   */
-  const continueApplication = (): void => {
-    useStore.setState({
-      saveToEmail: email,
-      path: ApplicationPath.SaveAndReturn,
-    });
-  };
-
-  const sendResumeEmailMutation = useMutation({
-    mutationFn: sendResumeEmail,
-    onSuccess: () => setPageStatus(Status.Success),
-    onError: () => setPageStatus(Status.Error),
-  });
-
-  const validateSessionMutation = useMutation({
-    mutationFn: validateSession,
-    onSuccess: (data) => {
-      const reconciledSessionData = data.reconciledSessionData;
-      setReconciliationResponse(data);
-      resumeSession(reconciledSessionData);
-      const shouldSkipReconciliation = data.message.match(
-        /Payment process initiated, skipping reconciliation/,
-      );
-
-      // Skip reconciliation page if applicant has started payment
-      shouldSkipReconciliation
-        ? continueApplication()
-        : setPageStatus(Status.Validated);
-    },
-    onError: (error) => {
-      if (isAxiosError(error)) {
-        if (error.response?.status === 403) {
-          const lockedSessionResponse = error.response
-            .data as LockedSessionResponse;
-          setPaymentRequest(lockedSessionResponse.paymentRequest);
-          setPageStatus(Status.LockedSession);
-          return;
-        }
-      }
-      console.debug(error);
-      setPageStatus(Status.InvalidSession);
-    },
-  });
-
-  /**
-   * Handle all submit "paths" that leads a user to this page
-   */
-  const handleSubmit = () => {
-    setPageStatus(Status.Validating);
-    sessionId
-      ? validateSessionMutation.mutate({
-          payload: { email, sessionId },
-        })
-      : sendResumeEmailMutation.mutate({
-          payload: { email, teamSlug },
-        });
-  };
-
-  /**
-   * Allow user to try again, with a different email address
-   */
-  const retryWithNewEmailAddress = () => {
-    setEmail("");
-    setPageStatus(Status.EmailRequired);
-  };
-
-  return {
-    [Status.EmailRequired]: <EmailRequired setEmail={setEmail} />,
-    [Status.Validating]: (
-      <DelayedLoadingIndicator
-        text={sessionId ? "Validating..." : "Sending..."}
-      />
-    ),
-    [Status.Validated]: (
-      <ValidationSuccess
-        reconciliationResponse={reconciliationResponse!}
-        continueApplication={continueApplication}
-      />
-    ),
-    [Status.InvalidSession]: (
-      <InvalidSession retry={retryWithNewEmailAddress} />
-    ),
-    [Status.LockedSession]: <LockedSession paymentRequest={paymentRequest} />,
-    [Status.Success]: <EmailSuccess />,
-    [Status.Error]: <EmailError retry={() => handleSubmit()} />,
-  }[pageStatus];
-};
-
-export default ResumePage;
